requirements.txt
Flask
Flask-SocketIO
eventlet
pymongo
python-dotenv
